+++ CHANGELOG.md
@@ -7,4 +7,9 @@
 - Feature (distinctive phonological features) structures
 - Phoneme builders for accent phoneme sets
 - General American English phoneme set
+- Syllable structure
+- Word structure
 
+### Changed
+- Phoneme symbol is char (not string)
+- Phoneme implements Copy